+++ resources/lang/cs/authority.php
@@ -24,7 +24,8 @@ return array(
     'back-to-artists'   => 'seznam autorů',
     'alternative_names' => 'příp.',
     'places'            => 'působení',
-    'links'             => 'externí odkazy',
+    'external_links'    => 'externí odkazy',
+    'source_links'      => 'použité zdroje',
     'relationships'     => 'vztahy',
 
     'artworks_by_artist' => 'díla autora',