+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
+### Added
+- admin form and view for 'sources' links
+
 ## [2.19.0] - 2021-02-18
 ### Fixed
 - paging issues with user collections